Bug 948302 - [Rocketbar] Entering URL launches browser frame r=daleharvey
