+++ apps/browser/index.html
@@ -34,6 +34,7 @@
     <script defer src="js/browser_db.js"></script>
     <script defer src="js/toolbar.js"></script>
     <script defer src="js/utilities.js"></script>
+    <script defer src="shared/js/url_helper.js"></script>
 
     <!-- Resources -->
     <link rel="resource" type="image/png" href="shared/resources/branding/Browser_60.png"/>
