+++ apps/search/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width,user-scalable=no,initial-scale=1.0">
     <link rel="stylesheet" href="style/search.css">
+    <script defer src="shared/js/url_helper.js"></script>
     <script defer src="js/search.js"></script>
     <script defer src="js/providers/contacts.js"></script>
     <script defer src="js/providers/eme.js"></script>
