+++ apps/search/manifest.webapp
@@ -20,6 +20,7 @@
   },
   "permissions": {
     "webapps-manage":{},
+    "open-remote-window":{},
     "contacts":{ "access": "readonly" }
   },
   "connections": {
