+++ apps/search/style/search.css
@@ -1,5 +1,4 @@
 body {
-  background: #000;
   color: #fff;
 }
