+++ apps/system/style/rocketbar.css
@@ -4,7 +4,6 @@
   left: 0;
   width: 100%;
   height: calc(100%);
-  background-color: #0B0B0C;
   pointer-events: none;
   transform: translateY(-100%);
   transition: transform 0.3s ease 0.8s;
@@ -40,6 +39,10 @@
 
 #search-bar .search-container {
   height: calc(100% - 6rem);
+}
+
+#search-container {
+  background-color: rgba(0, 0, 0, 0.5);
 }
 
 #search-bar .search-container iframe {
